+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - fix: Remove enum specifier for SentryLevel #822
+- feat: Send environment 'production' if nothing was set #825
 
 ## 6.0.6